+++ common/src/main/java/tocraft/walkers/traits/impl/CantFreezeTrait.java
@@ -1,0 +1,47 @@
+package tocraft.walkers.traits.impl;
+
+import com.mojang.serialization.MapCodec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.texture.TextureAtlasSprite;
+import net.minecraft.client.resources.model.BakedModel;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import org.jetbrains.annotations.Nullable;
+import tocraft.walkers.Walkers;
+import tocraft.walkers.traits.ShapeTrait;
+
+public class CantFreezeTrait<E extends LivingEntity> extends ShapeTrait<E> {
+    public static final ResourceLocation ID = Walkers.id("cant_freeze");
+    public static final MapCodec<CantFreezeTrait<?>> CODEC = RecordCodecBuilder.mapCodec((instance) -> instance.stable(new CantFreezeTrait<>()));
+
+    @Override
+    public ResourceLocation getId() {
+        return ID;
+    }
+
+    @Override
+    public MapCodec<? extends ShapeTrait<?>> codec() {
+        return CODEC;
+    }
+
+    @Override
+    public boolean canBeRegisteredMultipleTimes() {
+        return false;
+    }
+
+    @Override
+    @Environment(EnvType.CLIENT)
+    public @Nullable TextureAtlasSprite getIcon() {
+        BakedModel itemModel = Minecraft.getInstance().getItemRenderer().getModel(new ItemStack(Items.PACKED_ICE), null, null, 15);
+        if (itemModel != null) {
+            return itemModel.getParticleIcon();
+        } else {
+            return super.getIcon();
+        }
+    }
+}